@@ -29,7 +29,7 @@ use typst_library::model::{FootnoteElem, FootnoteEntry, LineNumberingScope, ParL
 use typst_library::pdf::ArtifactKind;
 use typst_library::routines::{Arenas, FragmentKind, Pair, RealizationKind, Routines};
 use typst_library::text::TextElem;
-use typst_utils::{NonZeroExt, Numeric};
+use typst_utils::{NonZeroExt, Numeric, Protected};
 
 use self::block::{layout_multi_block, layout_single_block};
 use self::collect::{
@@ -63,7 +63,7 @@ pub fn layout_fragment(
     layout_fragment_impl(
         engine.routines,
         engine.world,
-        engine.introspector,
+        engine.introspector.into_raw(),
         engine.traced,
         TrackedMut::reborrow_mut(&mut engine.sink),
         engine.route.track(),
@@ -91,7 +91,7 @@ pub fn layout_columns(
     layout_fragment_impl(
         engine.routines,
         engine.world,
-        engine.introspector,
+        engine.introspector.into_raw(),
         engine.traced,
         TrackedMut::reborrow_mut(&mut engine.sink),
         engine.route.track(),
@@ -128,6 +128,7 @@ fn layout_fragment_impl(
         bail!(content.span(), "cannot expand into infinite height");
     }
 
+    let introspector = Protected::from_raw(introspector);
     let link = LocatorLink::new(locator);
     let mut locator = Locator::link(&link).split();
     let mut engine = Engine {
