@@ -4,12 +4,13 @@ use std::ops::Deref;
 use std::sync::Arc;
 
 use az::SaturatingAs;
+use comemo::Tracked;
 use rustybuzz::{BufferFlags, Feature, ShapePlan, UnicodeBuffer};
 use ttf_parser::Tag;
 use ttf_parser::gsub::SubstitutionSubtable;
 use typst_library::World;
 use typst_library::engine::Engine;
-use typst_library::foundations::{Smart, StyleChain};
+use typst_library::foundations::{Regex, Smart, StyleChain};
 use typst_library::layout::{Abs, Dir, Em, Frame, FrameItem, Point, Rel, Size};
 use typst_library::model::{JustificationLimits, ParElem};
 use typst_library::text::{
@@ -790,7 +791,7 @@ fn shape<'a>(
     let size = styles.resolve(TextElem::size);
     let shift_settings = styles.get(TextElem::shift_settings);
     let mut ctx = ShapingContext {
-        engine,
+        world: engine.world,
         size,
         glyphs: vec![],
         used: vec![],
@@ -827,13 +828,9 @@ fn shape<'a>(
 }
 
 /// Holds shaping results and metadata common to all shaped segments.
-struct ShapingContext<'a, 'v> {
-    engine: &'a Engine<'v>,
+struct ShapingContext<'a> {
+    world: Tracked<'a, dyn World + 'a>,
     glyphs: Vec<ShapedGlyph>,
-    /// Font families that have been used with unlimited coverage.
-    ///
-    /// These font families are considered exhausted and will not be used again,
-    /// even if they are declared again (e.g., during fallback after normal selection).
     used: Vec<Font>,
     styles: StyleChain<'a>,
     size: Abs,
@@ -844,9 +841,98 @@ struct ShapingContext<'a, 'v> {
     shift_settings: Option<ShiftSettings>,
 }
 
+pub trait SharedShapingContext<'a> {
+    fn world(&self) -> Tracked<'a, dyn World + 'a>;
+
+    /// Font families that have been used with unlimited coverage.
+    ///
+    /// These font families are considered exhausted and will not be used again,
+    /// even if they are declared again (e.g., during fallback after normal selection).
+    fn used(&mut self) -> &mut Vec<Font>;
+
+    fn first(&self) -> Option<&Font>;
+
+    fn variant(&self) -> FontVariant;
+
+    fn fallback(&self) -> bool;
+}
+
+impl<'a> SharedShapingContext<'a> for ShapingContext<'a> {
+    fn world(&self) -> Tracked<'a, dyn World + 'a> {
+        self.world
+    }
+
+    fn used(&mut self) -> &mut Vec<Font> {
+        &mut self.used
+    }
+
+    fn first(&self) -> Option<&Font> {
+        self.used.first()
+    }
+
+    fn variant(&self) -> FontVariant {
+        self.variant
+    }
+
+    fn fallback(&self) -> bool {
+        self.fallback
+    }
+}
+
+pub fn get_font_and_covers<'a, C, F>(
+    ctx: &mut C,
+    text: &str,
+    mut families: impl Iterator<Item = &'a FontFamily>,
+    mut shape_tofus: F,
+) -> Option<(Font, Option<&'a Regex>)>
+where
+    C: SharedShapingContext<'a>,
+    F: FnMut(&mut C, &str, Font),
+{
+    // Find the next available family.
+    let world = ctx.world();
+    let book = world.book();
+    let mut selection = None;
+    let mut covers = None;
+    for family in families.by_ref() {
+        selection = book
+            .select(family.as_str(), ctx.variant())
+            .and_then(|id| world.font(id))
+            .filter(|font| !ctx.used().contains(font));
+        if selection.is_some() {
+            covers = family.covers();
+            break;
+        }
+    }
+
+    // Do font fallback if the families are exhausted and fallback is enabled.
+    if selection.is_none() && ctx.fallback() {
+        let first = ctx.first().map(Font::info);
+        selection = book
+            .select_fallback(first, ctx.variant(), text)
+            .and_then(|id| world.font(id))
+            .filter(|font| !ctx.used().contains(font));
+    }
+
+    // Extract the font id or shape notdef glyphs if we couldn't find any font.
+    let Some(font) = selection else {
+        if let Some(font) = ctx.used().first().cloned() {
+            shape_tofus(ctx, text, font);
+        }
+        return None;
+    };
+
+    // This font has been exhausted and will not be used again.
+    if covers.is_none() {
+        ctx.used().push(font.clone());
+    }
+
+    Some((font, covers))
+}
+
 /// Shape text with font fallback using the `families` iterator.
 fn shape_segment<'a>(
-    ctx: &mut ShapingContext,
+    ctx: &mut ShapingContext<'a>,
     base: usize,
     text: &str,
     mut families: impl Iterator<Item = &'a FontFamily> + Clone,
@@ -859,43 +945,13 @@ fn shape_segment<'a>(
         return;
     }
 
-    // Find the next available family.
-    let world = ctx.engine.world;
-    let book = world.book();
-    let mut selection = None;
-    let mut covers = None;
-    for family in families.by_ref() {
-        selection = book
-            .select(family.as_str(), ctx.variant)
-            .and_then(|id| world.font(id))
-            .filter(|font| !ctx.used.contains(font));
-        if selection.is_some() {
-            covers = family.covers();
-            break;
-        }
-    }
-
-    // Do font fallback if the families are exhausted and fallback is enabled.
-    if selection.is_none() && ctx.fallback {
-        let first = ctx.used.first().map(Font::info);
-        selection = book
-            .select_fallback(first, ctx.variant, text)
-            .and_then(|id| world.font(id))
-            .filter(|font| !ctx.used.contains(font));
-    }
-
-    // Extract the font id or shape notdef glyphs if we couldn't find any font.
-    let Some(font) = selection else {
-        if let Some(font) = ctx.used.first().cloned() {
+    let Some((font, covers)) =
+        get_font_and_covers(ctx, text, families.by_ref(), |ctx, text, font| {
             shape_tofus(ctx, base, text, font);
-        }
+        })
+    else {
         return;
     };
-
-    // This font has been exhausted and will not be used again.
-    if covers.is_none() {
-        ctx.used.push(font.clone());
-    }
 
     // Fill the buffer with our text.
     let mut buffer = UnicodeBuffer::new();
