@@ -21,8 +21,8 @@ use typst_utils::{ManuallyHash, NonZeroExt, PicoStr};
 
 use crate::World;
 use crate::diag::{
-    At, HintedStrResult, LoadError, LoadResult, LoadedWithin, ReportPos, SourceResult,
-    StrResult, bail, error, warning,
+    At, HintedStrResult, LoadError, LoadResult, LoadedWithin, ReportPos,
+    SourceDiagnostic, SourceResult, StrResult, bail, error, warning,
 };
 use crate::engine::{Engine, Sink};
 use crate::foundations::{
@@ -30,7 +30,9 @@ use crate::foundations::{
     OneOrMultiple, Packed, Reflect, Scope, ShowSet, Smart, StyleChain, Styles,
     Synthesize, Value, elem,
 };
-use crate::introspection::{Introspector, Locatable, Location};
+use crate::introspection::{
+    History, Introspect, Introspector, Locatable, Location, QueryIntrospection,
+};
 use crate::layout::{BlockBody, BlockElem, Em, HElem, PadElem};
 use crate::loading::{DataSource, Load, LoadSource, Loaded, format_yaml_error};
 use crate::model::{
@@ -151,9 +153,10 @@ pub struct BibliographyElem {
 
 impl BibliographyElem {
     /// Find the document's bibliography.
-    pub fn find(introspector: Tracked<Introspector>) -> StrResult<Packed<Self>> {
-        let query = introspector.query(&Self::ELEM.select());
-        let mut iter = query.iter();
+    pub fn find(engine: &mut Engine, span: Span) -> StrResult<Packed<Self>> {
+        let elems = engine.introspect(QueryIntrospection(Self::ELEM.select(), span));
+
+        let mut iter = elems.iter();
         let Some(elem) = iter.next() else {
             bail!("the document does not contain a bibliography");
         };
@@ -166,10 +169,9 @@ impl BibliographyElem {
     }
 
     /// Whether the bibliography contains the given key.
-    pub fn has(engine: &Engine, key: Label) -> bool {
+    pub fn has(engine: &mut Engine, key: Label, span: Span) -> bool {
         engine
-            .introspector
-            .query(&Self::ELEM.select())
+            .introspect(QueryIntrospection(Self::ELEM.select(), span))
             .iter()
             .any(|elem| elem.to_packed::<Self>().unwrap().sources.derived.has(key))
     }
@@ -556,8 +558,21 @@ pub struct Works {
 
 impl Works {
     /// Generate all citations and the whole bibliography.
-    pub fn generate(engine: &Engine) -> StrResult<Arc<Works>> {
-        Self::generate_impl(engine.routines, engine.world, engine.introspector)
+    pub fn generate(engine: &mut Engine, span: Span) -> SourceResult<Arc<Works>> {
+        let bibliography = BibliographyElem::find(engine, span).at(span)?;
+        let groups = engine.introspect(CiteGroupIntrospection(span));
+        Self::generate_impl(engine.routines, engine.world, bibliography, groups).at(span)
+    }
+
+    /// Generate all citations and the whole bibliography, given an existing
+    /// bibliography (no need to query it).
+    pub fn with_bibliography(
+        engine: &mut Engine,
+        bibliography: Packed<BibliographyElem>,
+    ) -> SourceResult<Arc<Works>> {
+        let span = bibliography.span();
+        let groups = engine.introspect(CiteGroupIntrospection(span));
+        Self::generate_impl(engine.routines, engine.world, bibliography, groups).at(span)
     }
 
     /// The internal implementation of [`Works::generate`].
@@ -565,9 +580,10 @@ impl Works {
     fn generate_impl(
         routines: &Routines,
         world: Tracked<dyn World + '_>,
-        introspector: Tracked<Introspector>,
+        bibliography: Packed<BibliographyElem>,
+        groups: EcoVec<Content>,
     ) -> StrResult<Arc<Works>> {
-        let mut generator = Generator::new(routines, world, introspector)?;
+        let mut generator = Generator::new(routines, world, bibliography, groups)?;
         let rendered = generator.drive();
         let works = generator.display(&rendered)?;
         Ok(Arc::new(works))
@@ -592,6 +608,34 @@ impl Works {
                 }
             })
             .at(elem.span())
+    }
+}
+
+/// Retrieves all citation groups in the document.
+///
+/// This is separate from `QueryIntrospection` so that we can customize the
+/// diagnostic as the `CiteGroup` is internal. The default query message is also
+/// not that helpful in this case.
+#[derive(Debug, Clone, PartialEq, Hash)]
+struct CiteGroupIntrospection(Span);
+
+impl Introspect for CiteGroupIntrospection {
+    type Output = EcoVec<Content>;
+
+    fn introspect(
+        &self,
+        _: &mut Engine,
+        introspector: Tracked<Introspector>,
+    ) -> Self::Output {
+        introspector.query(&CiteGroup::ELEM.select())
+    }
+
+    fn diagnose(&self, _: &History<Self::Output>) -> SourceDiagnostic {
+        warning!(
+            self.0, "citation grouping did not stabilize";
+            hint: "this can happen if the citations and bibliographies in the \
+                   document did not stabilize by the end of the third layout iteration"
+        )
     }
 }
 
@@ -641,10 +685,9 @@ impl<'a> Generator<'a> {
     fn new(
         routines: &'a Routines,
         world: Tracked<'a, dyn World + 'a>,
-        introspector: Tracked<Introspector>,
+        bibliography: Packed<BibliographyElem>,
+        groups: EcoVec<Content>,
     ) -> StrResult<Self> {
-        let bibliography = BibliographyElem::find(introspector)?;
-        let groups = introspector.query(&CiteGroup::ELEM.select());
         let infos = Vec::with_capacity(groups.len());
         Ok(Self {
             routines,
