@@ -4,14 +4,20 @@ use std::path::{Path, PathBuf};
 use std::str::FromStr;
 use std::sync::LazyLock;
 
-use bitflags::bitflags;
+use bitflags::{Flags, bitflags};
 use ecow::{EcoString, eco_format};
 use rustc_hash::{FxHashMap, FxHashSet};
+use typst::foundations::Bytes;
+use typst_pdf::PdfStandard;
 use typst_syntax::package::PackageVersion;
 use typst_syntax::{
     FileId, Lines, Source, VirtualPath, is_id_continue, is_ident, is_newline,
 };
 use unscanny::Scanner;
+
+use crate::output::HashedRefs;
+use crate::world::{read, system_path};
+use crate::{ARGS, REF_PATH, STORE_PATH, SUITE_PATH};
 
 /// Collects all tests from all files.
 ///
@@ -62,59 +68,230 @@ impl Display for FilePos {
 }
 
 bitflags! {
+    /// Just used for parsing attribute flags.
     #[derive(Copy, Clone)]
-    struct AttrFlags: u8 {
-        const RENDER = 1 << 0;
-        const HTML = 1 << 1;
-        const PDFTAGS = 1 << 2;
-        const LARGE = 1 << 3;
-        const NOPDFUA = 1 << 4;
+    struct AttrFlags: u16 {
+        const LARGE = 1 << 0;
     }
 }
 
-impl AttrFlags {
-    const NON_RENDER: Self = Self::HTML.union(Self::PDFTAGS);
-
-    pub fn targets(self) -> Targets {
-        let mut targets = Targets::empty();
-        if self.contains(Self::RENDER) || (self & Self::NON_RENDER).is_empty() {
-            targets |= Targets::RENDER;
-        }
-        if self.contains(Self::HTML) {
-            targets |= Targets::HTML;
-        }
-        if self.contains(Self::PDFTAGS) {
-            targets |= Targets::PDFTAGS;
-        }
-        targets
-    }
-}
-
+/// The parsed and evaluated test attributes specified in the test header.
 #[derive(Debug, Default, Copy, Clone, Eq, PartialEq)]
 pub struct Attrs {
     pub large: bool,
-    pub pdf_ua: bool,
-    pub targets: Targets,
+    pub pdf_standard: Option<PdfStandard>,
+    /// The test stages that are either directly specified or are implied by a
+    /// test attribute. If not specified otherwise by the `--stages` flag a
+    /// reference output will be generated.
+    pub stages: TestStages,
 }
+
+impl Attrs {
+    /// Whether the reference output should be compared and saved.
+    pub fn should_check_ref(&self, output: TestOutput) -> bool {
+        // TODO: Enable PDF and SVG once we have a diffing tool for hashed references.
+        ARGS.should_run(self.stages & output.into())
+            && output != TestOutput::Pdf
+            && output != TestOutput::Svg
+    }
+}
+
+pub trait TestStage: Into<TestStages> + Display + Copy {}
 
 bitflags! {
+    /// The stages a test in ran through. This combines both compilation targets
+    /// and output formats.
     #[derive(Debug, Default, Copy, Clone, Eq, PartialEq)]
-    pub struct Targets: u8 {
-        const RENDER = 0x1;
-        const HTML = 0x2;
-        const PDFTAGS = 0x4;
+    pub struct TestStages: u8 {
+        const PAGED = 1 << 0;
+        const RENDER = 1 << 1;
+        const PDF = 1 << 2;
+        const PDFTAGS = 1 << 3;
+        const SVG = 1 << 4;
+        const HTML = 1 << 5;
     }
 }
 
-impl Targets {
-    pub fn from_file_extension(ext: &str) -> Option<Self> {
-        Some(match ext {
-            "png" => Self::RENDER,
-            "html" => Self::HTML,
-            "yml" => Self::PDFTAGS,
-            _ => return None,
+macro_rules! union {
+    ($union:expr) => {
+        $union
+    };
+    ($a:expr, $b:expr $(,$flag:expr)*$(,)?) => {
+        union!($a.union($b) $(,$flag)*)
+    };
+}
+
+impl TestStages {
+    /// All stages that require the paged target.
+    pub const PAGED_STAGES: Self = union!(
+        TestStages::PAGED,
+        TestStages::RENDER,
+        TestStages::PDF,
+        TestStages::PDFTAGS,
+        TestStages::SVG,
+    );
+
+    /// All stages that require a pdf document.
+    pub const PDF_STAGES: Self = union!(TestStages::PDF, TestStages::PDFTAGS);
+
+    /// The union the supplied stages and their implied stages.
+    ///
+    /// The `paged` target will test `render`, `pdf`, and `svg` by default.
+    pub fn with_implied(&self) -> TestStages {
+        let mut res = *self;
+        for flag in self.iter() {
+            res |= bitflags::bitflags_match!(flag, {
+                TestStages::PAGED => TestStages::RENDER | TestStages::PDF | TestStages::SVG,
+                TestStages::RENDER => TestStages::empty(),
+                TestStages::PDF => TestStages::empty(),
+                TestStages::PDFTAGS => TestStages::empty(),
+                TestStages::SVG => TestStages::empty(),
+                TestStages::HTML => TestStages::empty(),
+                _ => unreachable!(),
+            });
+        }
+        res
+    }
+}
+
+impl Display for TestStages {
+    fn fmt(&self, f: &mut Formatter<'_>) -> fmt::Result {
+        for (i, flag) in self.iter().enumerate() {
+            if i != 0 {
+                f.write_str(", ")?;
+            }
+            bitflags::bitflags_match!(flag, {
+                TestStages::PAGED => Display::fmt(&TestTarget::Paged, f),
+                TestStages::RENDER => Display::fmt(&TestOutput::Render, f),
+                TestStages::PDF => Display::fmt(&TestOutput::Pdf, f),
+                TestStages::PDFTAGS => Display::fmt(&TestOutput::Pdftags, f),
+                TestStages::SVG => Display::fmt(&TestOutput::Svg, f),
+                TestStages::HTML => Display::fmt(&TestTarget::Html, f),
+                _ => unreachable!(),
+            })?;
+        }
+        Ok(())
+    }
+}
+
+/// A compilation target, analog to [`typst::Target`].
+#[derive(Debug, Copy, Clone, Eq, PartialEq)]
+#[repr(u8)]
+pub enum TestTarget {
+    Paged = TestStages::PAGED.bits(),
+    Html = TestStages::HTML.bits(),
+}
+
+impl TestStage for TestTarget {}
+
+impl From<TestTarget> for TestStages {
+    fn from(value: TestTarget) -> Self {
+        TestStages::from_bits(value as u8).unwrap()
+    }
+}
+
+impl Display for TestTarget {
+    fn fmt(&self, f: &mut Formatter<'_>) -> fmt::Result {
+        f.write_str(match self {
+            TestTarget::Paged => "paged",
+            TestTarget::Html => "html",
         })
     }
+}
+
+/// A test output format.
+#[derive(Debug, Copy, Clone, Eq, PartialEq, Hash)]
+#[repr(u8)]
+pub enum TestOutput {
+    Render = TestStages::RENDER.bits(),
+    Pdf = TestStages::PDF.bits(),
+    Pdftags = TestStages::PDFTAGS.bits(),
+    Svg = TestStages::SVG.bits(),
+    Html = TestStages::HTML.bits(),
+}
+
+impl TestOutput {
+    pub const ALL: [Self; 5] =
+        [Self::Render, Self::Pdf, Self::Pdftags, Self::Svg, Self::Html];
+
+    fn from_sub_dir(dir: &str) -> Option<Self> {
+        Self::ALL.into_iter().find(|o| o.sub_dir() == dir)
+    }
+
+    /// The sub directory inside the [`REF_PATH`] and [`STORE_PATH`].
+    pub const fn sub_dir(&self) -> &'static str {
+        match self {
+            Self::Render => "render",
+            Self::Pdf => "pdf",
+            Self::Pdftags => "pdftags",
+            Self::Svg => "svg",
+            Self::Html => "html",
+        }
+    }
+
+    /// The file extension used for live outputs and file references.
+    pub const fn extension(&self) -> &'static str {
+        match self {
+            Self::Render => "png",
+            Self::Pdf => "pdf",
+            Self::Pdftags => "yml",
+            Self::Svg => "svg",
+            Self::Html => "html",
+        }
+    }
+
+    /// The path at which the live output will be stored for inspection.
+    pub fn live_path(&self, name: &str) -> PathBuf {
+        let dir = self.sub_dir();
+        let ext = self.extension();
+        PathBuf::from(format!("{STORE_PATH}/{dir}/{name}.{ext}"))
+    }
+
+    /// The path at which file references will be saved.
+    pub fn file_ref_path(&self, name: &str) -> PathBuf {
+        let dir = self.sub_dir();
+        let ext = self.extension();
+        PathBuf::from(format!("{REF_PATH}/{dir}/{name}.{ext}"))
+    }
+
+    /// The path at which hashed references will be saved.
+    pub fn hashed_ref_path(self, source_path: &Path) -> PathBuf {
+        let sub_dir = self.sub_dir();
+        let sub_path = source_path.strip_prefix(SUITE_PATH).unwrap();
+        let trimmed_path = sub_path.to_str().unwrap().strip_suffix(".typ");
+        let file_name = trimmed_path.unwrap().replace("/", "-");
+        PathBuf::from(format!("{REF_PATH}/{sub_dir}/{file_name}.txt"))
+    }
+
+    /// The output kind.
+    fn kind(&self) -> TestOutputKind {
+        match self {
+            TestOutput::Render | TestOutput::Pdftags | TestOutput::Html => {
+                TestOutputKind::File
+            }
+            TestOutput::Pdf | TestOutput::Svg => TestOutputKind::Hash,
+        }
+    }
+}
+
+impl TestStage for TestOutput {}
+
+impl From<TestOutput> for TestStages {
+    fn from(value: TestOutput) -> Self {
+        TestStages::from_bits(value as u8).unwrap()
+    }
+}
+
+impl Display for TestOutput {
+    fn fmt(&self, f: &mut Formatter<'_>) -> fmt::Result {
+        f.write_str(self.sub_dir())
+    }
+}
+
+/// Whether the output format produces hashed or file references.
+enum TestOutputKind {
+    Hash,
+    File,
 }
 
 /// The size of a file.
@@ -216,46 +393,113 @@ impl Collector {
         }
     }
 
-    /// Walks through all reference output and ensures that a test exists for
-    /// each one.
+    /// Walks through all reference outputs and ensures that a matching test
+    /// exists.
     fn walk_references(&mut self) {
         for entry in walkdir::WalkDir::new(crate::REF_PATH).sort_by_file_name() {
             let entry = entry.unwrap();
-            let path = entry.path();
-            let Some(file_target) = path.extension().and_then(|ext| {
-                let str = ext.to_str()?;
-                Targets::from_file_extension(str)
-            }) else {
+            if !entry.file_type().is_file() {
                 continue;
-            };
-
-            let stem = path.file_stem().unwrap().to_string_lossy();
-            let name = &*stem;
-
-            let Some((pos, attrs)) = self.seen.get(name) else {
-                self.errors.push(TestParseError {
-                    pos: FilePos::new(path, 0),
-                    message: "dangling reference output".into(),
-                });
-                continue;
-            };
-
-            if !attrs.targets.contains(file_target) {
-                self.errors.push(TestParseError {
-                    pos: FilePos::new(path, 0),
-                    message: "dangling reference output".into(),
-                });
             }
 
-            let len = path.metadata().unwrap().len() as usize;
-            if !attrs.large && len > crate::REF_LIMIT {
+            let path = entry.path();
+            let sub_path = path.strip_prefix(crate::REF_PATH).unwrap();
+            let sub_dir = sub_path.components().next().unwrap();
+            let Some(output) =
+                TestOutput::from_sub_dir(sub_dir.as_os_str().to_str().unwrap())
+            else {
+                continue;
+            };
+
+            match output.kind() {
+                TestOutputKind::File => self.check_dangling_file_references(path, output),
+                TestOutputKind::Hash => {
+                    self.check_dangling_hashed_references(path, output)
+                }
+            }
+        }
+    }
+
+    fn check_dangling_file_references(&mut self, path: &Path, output: TestOutput) {
+        let stem = path.file_stem().unwrap().to_string_lossy();
+        let name = &*stem;
+
+        let Some((pos, attrs)) = self.seen.get(name) else {
+            self.errors.push(TestParseError {
+                pos: FilePos::new(path, 0),
+                message: "dangling reference output".into(),
+            });
+            return;
+        };
+
+        if !attrs.stages.contains(output.into()) {
+            self.errors.push(TestParseError {
+                pos: FilePos::new(path, 0),
+                message: "dangling reference output".into(),
+            });
+        }
+
+        let len = path.metadata().unwrap().len() as usize;
+        if !attrs.large && len > crate::REF_LIMIT {
+            self.errors.push(TestParseError {
+                pos: pos.clone(),
+                message: format!(
+                    "reference output size exceeds {}, but the test is not marked as `large`",
+                    FileSize(crate::REF_LIMIT),
+                ),
+            });
+        }
+    }
+
+    fn check_dangling_hashed_references(&mut self, path: &Path, output: TestOutput) {
+        let string = std::fs::read_to_string(path).unwrap_or_default();
+        let Ok(hashed_refs) = HashedRefs::from_str(&string) else { return };
+        if hashed_refs.is_empty() {
+            self.errors.push(TestParseError {
+                pos: FilePos::new(path, 0),
+                message: "dangling empty reference hash file".into(),
+            });
+        }
+
+        let mut right_file = 0;
+        let mut wrong_file = Vec::new();
+        for (line, name) in hashed_refs.names().enumerate() {
+            let Some((pos, attrs)) = self.seen.get(name) else {
                 self.errors.push(TestParseError {
-                    pos: pos.clone(),
-                    message: format!(
-                        "reference output size exceeds {}, but the test is not marked as `large`",
-                        FileSize(crate::REF_LIMIT),
-                    ),
+                    pos: FilePos::new(path, line),
+                    message: format!("dangling reference hash ({name})"),
                 });
+                continue;
+            };
+
+            if !attrs.stages.contains(output.into()) {
+                self.errors.push(TestParseError {
+                    pos: FilePos::new(path, line),
+                    message: format!("dangling reference hash ({name})"),
+                });
+                continue;
+            }
+
+            if output.hashed_ref_path(&pos.path) == path {
+                right_file += 1;
+            } else {
+                wrong_file.push((line, name));
+            }
+        }
+
+        if !wrong_file.is_empty() {
+            if right_file == 0 {
+                self.errors.push(TestParseError {
+                    pos: FilePos::new(path, 0),
+                    message: "dangling reference hash file".into(),
+                });
+            } else {
+                for (line, name) in wrong_file {
+                    self.errors.push(TestParseError {
+                        pos: FilePos::new(path, line),
+                        message: format!("dangling reference hash ({name})"),
+                    });
+                }
             }
         }
     }
@@ -329,7 +573,9 @@ impl<'a> Parser<'a> {
 
             let text = self.s.from(start);
 
-            if !selected(&name, self.path.canonicalize().unwrap()) {
+            if !ARGS.should_run(attrs.stages)
+                || !selected(&name, self.path.canonicalize().unwrap())
+            {
                 self.collector.skipped += 1;
                 continue;
             }
@@ -356,35 +602,73 @@ impl<'a> Parser<'a> {
         }
     }
 
+    /// Parse the test attributes inside a test header.
     fn parse_attrs(&mut self) -> Attrs {
-        let mut parsed = AttrFlags::empty();
+        let mut stages = TestStages::empty();
+        let mut flags = AttrFlags::empty();
+        let mut pdf_standard = None;
         while !self.s.eat_if("---") {
-            let attr = self.s.eat_until(char::is_whitespace);
-            let flag = match attr {
-                "large" => AttrFlags::LARGE,
-                "html" => AttrFlags::HTML,
-                "render" => AttrFlags::RENDER,
-                "pdftags" => AttrFlags::PDFTAGS,
-                "nopdfua" => AttrFlags::NOPDFUA,
+            let attr_name = self.s.eat_while(is_id_continue);
+            let mut attr_params = None;
+            if self.s.eat_if('(') {
+                attr_params = Some(self.s.eat_until(')'));
+                if !self.s.eat_if(')') {
+                    self.error("expected closing parenthesis");
+                }
+            }
+            if !self.s.at(' ') {
+                self.error("expected a space after an attribute");
+            }
+
+            match attr_name {
+                "paged" => self.set_attr(attr_name, &mut stages, TestStages::PAGED),
+                "pdftags" => self.set_attr(attr_name, &mut stages, TestStages::PDFTAGS),
+                "pdfstandard" => {
+                    let Some(param) = attr_params.take() else {
+                        self.error("expected parameter for `pdfstandard`");
+                        continue;
+                    };
+                    pdf_standard = serde_yaml::from_str(param)
+                        .inspect_err(|e| {
+                            self.error(format!("unknown pdf standard `{param}`: {e}"))
+                        })
+                        .ok();
+                }
+                "html" => self.set_attr(attr_name, &mut stages, TestStages::HTML),
+                "large" => self.set_attr(attr_name, &mut flags, AttrFlags::LARGE),
+
                 found => {
                     self.error(format!(
                         "expected attribute or closing ---, found `{found}`"
                     ));
                     break;
                 }
-            };
-            if parsed.contains(flag) {
-                self.error(format!("duplicate attribute `{attr}`"));
             }
-            parsed.insert(flag);
+
+            if attr_params.is_some() {
+                self.error("unexpected attribute parameters");
+            }
+
             self.s.eat_while(' ');
         }
 
-        Attrs {
-            large: parsed.contains(AttrFlags::LARGE),
-            pdf_ua: !parsed.contains(AttrFlags::NOPDFUA),
-            targets: parsed.targets(),
+        if stages.is_empty() {
+            self.error("tests must specify at least one target or output");
         }
+
+        Attrs {
+            large: flags.contains(AttrFlags::LARGE),
+            pdf_standard,
+            stages: stages.with_implied(),
+        }
+    }
+
+    /// Set an attribute flag and check for duplicates.
+    fn set_attr<F: Flags + Copy>(&mut self, attr: &str, flags: &mut F, flag: F) {
+        if flags.contains(flag) {
+            self.error(format!("duplicate attribute `{attr}`"));
+        }
+        flags.insert(flag);
     }
 
     /// Skips the preamble of a test.
@@ -444,7 +728,8 @@ impl<'a> Parser<'a> {
             .s
             .eat_until(is_newline)
             .trim()
-            .replace("VERSION", &eco_format!("{}", PackageVersion::compiler()));
+            .replace("VERSION", &eco_format!("{}", PackageVersion::compiler()))
+            .replace("\\n", "\n");
 
         Some(Note {
             pos: FilePos::new(self.path, self.line),
@@ -455,19 +740,9 @@ impl<'a> Parser<'a> {
         })
     }
 
-    #[cfg(not(feature = "default"))]
-    fn parse_range_external(&mut self, _file: FileId) -> Option<Range<usize>> {
-        panic!("external file ranges are not expected when testing `typst_syntax`");
-    }
-
     /// Parse a range in an external file, optionally abbreviated as just a position
     /// if the range is empty.
-    #[cfg(feature = "default")]
     fn parse_range_external(&mut self, file: FileId) -> Option<Range<usize>> {
-        use typst::foundations::Bytes;
-
-        use crate::world::{read, system_path};
-
         let path = match system_path(file) {
             Ok(path) => path,
             Err(err) => {
