@@ -1,16 +1,12 @@
 //! Typst's test runner.
 
-#![cfg_attr(not(feature = "default"), allow(dead_code, unused_imports))]
-
 mod args;
 mod collect;
-mod logger;
-
-#[cfg(feature = "default")]
 mod custom;
-#[cfg(feature = "default")]
+mod logger;
+mod output;
+mod pdftags;
 mod run;
-#[cfg(feature = "default")]
 mod world;
 
 use std::path::{Path, PathBuf};
@@ -18,8 +14,9 @@ use std::sync::LazyLock;
 use std::time::Duration;
 
 use clap::Parser;
-use parking_lot::Mutex;
+use parking_lot::{Mutex, RwLock};
 use rayon::iter::{ParallelBridge, ParallelIterator};
+use rustc_hash::FxHashMap;
 
 use crate::args::{CliArguments, Command};
 use crate::collect::Test;
@@ -102,17 +99,14 @@ fn test() {
     }
 
     let parser_dirs = ARGS.parser_compare.clone().map(create_syntax_store);
-    #[cfg(not(feature = "default"))]
-    let parser_dirs = parser_dirs.or_else(|| Some(create_syntax_store(None)));
+
+    let hashes: [_; 2] = std::array::from_fn(|_| RwLock::new(FxHashMap::default()));
 
     let runner = |test: &Test| {
         if let Some((live_path, ref_path)) = &parser_dirs {
             run_parser_test(test, live_path, ref_path)
         } else {
-            #[cfg(feature = "default")]
-            return run::run(test);
-            #[cfg(not(feature = "default"))]
-            unreachable!();
+            run::run(&hashes, test)
         }
     };
 
@@ -139,12 +133,19 @@ fn test() {
         // to `typst::utils::Deferred` yielding.
         tests.iter().par_bridge().for_each(|test| {
             logger.lock().start(test);
-            let result = std::panic::catch_unwind(|| runner(test));
+
+            // This is in fact not formally unwind safe, but the code paths that
+            // hold a lock of the hashes are quite short and shouldn't panic.
+            let closure = std::panic::AssertUnwindSafe(|| runner(test));
+            let result = std::panic::catch_unwind(closure);
             logger.lock().end(test, result);
         });
 
         sender.send(()).unwrap();
     });
+
+    run::update_hash_refs::<output::Pdf>(&hashes);
+    run::update_hash_refs::<output::Svg>(&hashes);
 
     let passed = logger.into_inner().finish();
     if !passed {
